Show participants who did not finish the experiment

Show information on the people who did not finish the experiment at (a) certain page(s) and/or app(s).

Usage

show_dropouts(oTree, final_apps = NULL, final_pages = NULL, saved_vars = NULL)

Arguments

oTree

A list of data frames that were created by import_otree.

final_apps

Character. The name(s) of the app(s) at which the participants have to finish the experiment.

final_pages

Character. The name(s) of the page(s) at which the participants have to finish the experiment.

saved_vars

The name(s) of variable(s) that need(s) to be shown in the list of information on dropout cases.

Value

This function returns a list of information on participants who did not finish the experiment.

In this list, you can find the following information:

- $full = A data frame that contains information on all participants who did not finish the study; it shows their participant codes, the names of the apps in which they left the experiment, the names of the pages in which they left the experiment, the names of the app data frames in which this information was found, and the dropout reason ("ENC", experiment not completed, combined with the name of the data frame in which the dropout was observed). Because participants usually appear in multiple app data frames, the $full data frame may contain several entries for each person.

- $unique = A data frame that contains similar information as the $full data frame but with only one row per participant and no information on the data frame in which the dropout was observed.

- $all_end = A table that provides information on the app and page combinations where participants ended the experiment. This table also includes information on participants who did not drop out of the experiment. The $all_end table is only shown if an $all_apps_wide data frame exists.

- $codes = A vector containing the participant codes of all participants who did not finish the experiment.

- $count = The number of all participants who did not finish the experiment.

It is important to note that if only the argument final_pages is set, this function does not distinguish between page names that reoccur in different apps.

If the columns end_app and end_page in the output are empty, these variables were not saved by oTree for the specific participants. This could be because empty rows were not deleted. This can be done by using the argument del_empty = TRUE" when using import_otree.

Examples

# Use package-internal list of oTree data frames
oTree <- gmoTree::oTree

# Show everyone who did not finish with the app "survey"
show_dropouts(oTree, final_apps = "survey")
#> $full
#>    participant.code session.code  end_app        end_page
#> 1          164r1hs4     7bfqtokx  chatapp            chat
#> 2          164r1hs4     7bfqtokx  chatapp            chat
#> 3          164r1hs4     7bfqtokx  chatapp            chat
#> 4          164r1hs4     7bfqtokx  chatapp            chat
#> 5          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 6          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 7          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 8          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 9          7wa8kk3d     l8ecm1uh dictator    Introduction
#> 10         7wa8kk3d     l8ecm1uh dictator    Introduction
#> 11         7wa8kk3d     l8ecm1uh dictator    Introduction
#> 12         7wa8kk3d     l8ecm1uh dictator    Introduction
#> 13         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 14         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 15         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 16         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 17         d4sq7zio     l8ecm1uh dictator    Introduction
#> 18         d4sq7zio     l8ecm1uh dictator    Introduction
#> 19         d4sq7zio     l8ecm1uh dictator    Introduction
#> 20         d4sq7zio     l8ecm1uh dictator    Introduction
#> 21         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 22         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 23         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 24         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 25         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 26         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 27         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 28         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 29         znri6myc     l8ecm1uh dictator    Introduction
#> 30         znri6myc     l8ecm1uh dictator    Introduction
#> 31         znri6myc     l8ecm1uh dictator    Introduction
#> 32         znri6myc     l8ecm1uh dictator    Introduction
#>                                                 reason
#> 1  Experiment not completed. Noticed at: all_apps_wide
#> 2        Experiment not completed. Noticed at: chatapp
#> 3       Experiment not completed. Noticed at: dictator
#> 4         Experiment not completed. Noticed at: survey
#> 5  Experiment not completed. Noticed at: all_apps_wide
#> 6        Experiment not completed. Noticed at: chatapp
#> 7       Experiment not completed. Noticed at: dictator
#> 8         Experiment not completed. Noticed at: survey
#> 9  Experiment not completed. Noticed at: all_apps_wide
#> 10       Experiment not completed. Noticed at: chatapp
#> 11      Experiment not completed. Noticed at: dictator
#> 12        Experiment not completed. Noticed at: survey
#> 13 Experiment not completed. Noticed at: all_apps_wide
#> 14       Experiment not completed. Noticed at: chatapp
#> 15      Experiment not completed. Noticed at: dictator
#> 16        Experiment not completed. Noticed at: survey
#> 17 Experiment not completed. Noticed at: all_apps_wide
#> 18       Experiment not completed. Noticed at: chatapp
#> 19      Experiment not completed. Noticed at: dictator
#> 20        Experiment not completed. Noticed at: survey
#> 21 Experiment not completed. Noticed at: all_apps_wide
#> 22       Experiment not completed. Noticed at: chatapp
#> 23      Experiment not completed. Noticed at: dictator
#> 24        Experiment not completed. Noticed at: survey
#> 25 Experiment not completed. Noticed at: all_apps_wide
#> 26       Experiment not completed. Noticed at: chatapp
#> 27      Experiment not completed. Noticed at: dictator
#> 28        Experiment not completed. Noticed at: survey
#> 29 Experiment not completed. Noticed at: all_apps_wide
#> 30       Experiment not completed. Noticed at: chatapp
#> 31      Experiment not completed. Noticed at: dictator
#> 32        Experiment not completed. Noticed at: survey
#> 
#> $unique
#>    participant.code session.code  end_app        end_page
#> 1          164r1hs4     7bfqtokx  chatapp            chat
#> 5          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 9          7wa8kk3d     l8ecm1uh dictator    Introduction
#> 13         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 17         d4sq7zio     l8ecm1uh dictator    Introduction
#> 21         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 25         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 29         znri6myc     l8ecm1uh dictator    Introduction
#>                      reason
#> 1  Experiment not completed
#> 5  Experiment not completed
#> 9  Experiment not completed
#> 13 Experiment not completed
#> 17 Experiment not completed
#> 21 Experiment not completed
#> 25 Experiment not completed
#> 29 Experiment not completed
#> 
#> $all_end
#>           
#>            Demographics Introduction Offer ResultsWaitPage chat
#>   chatapp             0            0     0               0    1
#>   dictator            0            4     0               3    0
#>   survey             20            0     1               0    0
#> 
#> $codes
#> [1] "164r1hs4" "2scvem7a" "7wa8kk3d" "ao9kqvqn" "d4sq7zio" "ktjz5jli" "vbhvhozv"
#> [8] "znri6myc"
#> 
#> $count
#> [1] 8
#> 

# Show everyone who did not finish with the page "Demographics"
show_dropouts(oTree, final_pages = "Demographics")
#> $full
#>    participant.code session.code  end_app        end_page
#> 1          164r1hs4     7bfqtokx  chatapp            chat
#> 2          164r1hs4     7bfqtokx  chatapp            chat
#> 3          164r1hs4     7bfqtokx  chatapp            chat
#> 4          164r1hs4     7bfqtokx  chatapp            chat
#> 5          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 6          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 7          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 8          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 9          7wa8kk3d     l8ecm1uh dictator    Introduction
#> 10         7wa8kk3d     l8ecm1uh dictator    Introduction
#> 11         7wa8kk3d     l8ecm1uh dictator    Introduction
#> 12         7wa8kk3d     l8ecm1uh dictator    Introduction
#> 13         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 14         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 15         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 16         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 17         d4sq7zio     l8ecm1uh dictator    Introduction
#> 18         d4sq7zio     l8ecm1uh dictator    Introduction
#> 19         d4sq7zio     l8ecm1uh dictator    Introduction
#> 20         d4sq7zio     l8ecm1uh dictator    Introduction
#> 21         jxh15obl     t0rog7nz   survey           Offer
#> 22         jxh15obl     t0rog7nz   survey           Offer
#> 23         jxh15obl     t0rog7nz   survey           Offer
#> 24         jxh15obl     t0rog7nz   survey           Offer
#> 25         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 26         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 27         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 28         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 29         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 30         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 31         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 32         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 33         znri6myc     l8ecm1uh dictator    Introduction
#> 34         znri6myc     l8ecm1uh dictator    Introduction
#> 35         znri6myc     l8ecm1uh dictator    Introduction
#> 36         znri6myc     l8ecm1uh dictator    Introduction
#>                                                 reason
#> 1  Experiment not completed. Noticed at: all_apps_wide
#> 2         Experiment not completed. Noticed at: survey
#> 3       Experiment not completed. Noticed at: dictator
#> 4        Experiment not completed. Noticed at: chatapp
#> 5         Experiment not completed. Noticed at: survey
#> 6       Experiment not completed. Noticed at: dictator
#> 7        Experiment not completed. Noticed at: chatapp
#> 8  Experiment not completed. Noticed at: all_apps_wide
#> 9         Experiment not completed. Noticed at: survey
#> 10      Experiment not completed. Noticed at: dictator
#> 11       Experiment not completed. Noticed at: chatapp
#> 12 Experiment not completed. Noticed at: all_apps_wide
#> 13        Experiment not completed. Noticed at: survey
#> 14      Experiment not completed. Noticed at: dictator
#> 15       Experiment not completed. Noticed at: chatapp
#> 16 Experiment not completed. Noticed at: all_apps_wide
#> 17      Experiment not completed. Noticed at: dictator
#> 18       Experiment not completed. Noticed at: chatapp
#> 19 Experiment not completed. Noticed at: all_apps_wide
#> 20        Experiment not completed. Noticed at: survey
#> 21       Experiment not completed. Noticed at: chatapp
#> 22 Experiment not completed. Noticed at: all_apps_wide
#> 23        Experiment not completed. Noticed at: survey
#> 24      Experiment not completed. Noticed at: dictator
#> 25        Experiment not completed. Noticed at: survey
#> 26      Experiment not completed. Noticed at: dictator
#> 27       Experiment not completed. Noticed at: chatapp
#> 28 Experiment not completed. Noticed at: all_apps_wide
#> 29        Experiment not completed. Noticed at: survey
#> 30      Experiment not completed. Noticed at: dictator
#> 31       Experiment not completed. Noticed at: chatapp
#> 32 Experiment not completed. Noticed at: all_apps_wide
#> 33        Experiment not completed. Noticed at: survey
#> 34      Experiment not completed. Noticed at: dictator
#> 35       Experiment not completed. Noticed at: chatapp
#> 36 Experiment not completed. Noticed at: all_apps_wide
#> 
#> $unique
#>    participant.code session.code  end_app        end_page
#> 1          164r1hs4     7bfqtokx  chatapp            chat
#> 5          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 9          7wa8kk3d     l8ecm1uh dictator    Introduction
#> 13         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 17         d4sq7zio     l8ecm1uh dictator    Introduction
#> 21         jxh15obl     t0rog7nz   survey           Offer
#> 25         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 29         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 33         znri6myc     l8ecm1uh dictator    Introduction
#>                      reason
#> 1  Experiment not completed
#> 5  Experiment not completed
#> 9  Experiment not completed
#> 13 Experiment not completed
#> 17 Experiment not completed
#> 21 Experiment not completed
#> 25 Experiment not completed
#> 29 Experiment not completed
#> 33 Experiment not completed
#> 
#> $all_end
#>           
#>            Demographics Introduction Offer ResultsWaitPage chat
#>   chatapp             0            0     0               0    1
#>   dictator            0            4     0               3    0
#>   survey             20            0     1               0    0
#> 
#> $codes
#> [1] "164r1hs4" "2scvem7a" "7wa8kk3d" "ao9kqvqn" "d4sq7zio" "jxh15obl" "ktjz5jli"
#> [8] "vbhvhozv" "znri6myc"
#> 
#> $count
#> [1] 9
#> 

# Show everyone who finished with the following apps: "survey," "dictator"
final_apps <- unique(oTree$all_apps_wide$participant._current_app_name)
final_apps <- final_apps[final_apps != "survey"]
final_apps <- final_apps[final_apps != "dictator"]
show_dropouts(oTree, final_apps = final_apps)
#> $full
#>     participant.code session.code  end_app        end_page
#> 1           1l5kal0r     t0rog7nz   survey    Demographics
#> 2           1l5kal0r     t0rog7nz   survey    Demographics
#> 3           1l5kal0r     t0rog7nz   survey    Demographics
#> 4           1l5kal0r     t0rog7nz   survey    Demographics
#> 5           1mn1kzs7     gii59rga   survey    Demographics
#> 6           2d72mfgh     7bfqtokx   survey    Demographics
#> 7           2d72mfgh     7bfqtokx   survey    Demographics
#> 8           2d72mfgh     7bfqtokx   survey    Demographics
#> 9           2d72mfgh     7bfqtokx   survey    Demographics
#> 10          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 11          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 12          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 13          2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 14          3ttf7yix     7bfqtokx   survey    Demographics
#> 15          3ttf7yix     7bfqtokx   survey    Demographics
#> 16          3ttf7yix     7bfqtokx   survey    Demographics
#> 17          3ttf7yix     7bfqtokx   survey    Demographics
#> 18          46kxib6w     7bfqtokx   survey    Demographics
#> 19          46kxib6w     7bfqtokx   survey    Demographics
#> 20          46kxib6w     7bfqtokx   survey    Demographics
#> 21          46kxib6w     7bfqtokx   survey    Demographics
#> 22          4zhzdmzo     7bfqtokx   survey    Demographics
#> 23          4zhzdmzo     7bfqtokx   survey    Demographics
#> 24          4zhzdmzo     7bfqtokx   survey    Demographics
#> 25          4zhzdmzo     7bfqtokx   survey    Demographics
#> 26          6cim3a5i     gii59rga   survey    Demographics
#> 27          7l0hmpcq     qyu3qg0q   survey    Demographics
#> 28          7l0hmpcq     qyu3qg0q   survey    Demographics
#> 29          7l0hmpcq     qyu3qg0q   survey    Demographics
#> 30          7l0hmpcq     qyu3qg0q   survey    Demographics
#> 31          7wa8kk3d     l8ecm1uh dictator    Introduction
#> 32          7wa8kk3d     l8ecm1uh dictator    Introduction
#> 33          7wa8kk3d     l8ecm1uh dictator    Introduction
#> 34          7wa8kk3d     l8ecm1uh dictator    Introduction
#> 35          a7dppel1     t0rog7nz   survey    Demographics
#> 36          a7dppel1     t0rog7nz   survey    Demographics
#> 37          a7dppel1     t0rog7nz   survey    Demographics
#> 38          a7dppel1     t0rog7nz   survey    Demographics
#> 39          ao9kqvqn     l8ecm1uh dictator    Introduction
#> 40          ao9kqvqn     l8ecm1uh dictator    Introduction
#> 41          ao9kqvqn     l8ecm1uh dictator    Introduction
#> 42          ao9kqvqn     l8ecm1uh dictator    Introduction
#> 43          byxco6zo     gii59rga   survey    Demographics
#> 44          d4sq7zio     l8ecm1uh dictator    Introduction
#> 45          d4sq7zio     l8ecm1uh dictator    Introduction
#> 46          d4sq7zio     l8ecm1uh dictator    Introduction
#> 47          d4sq7zio     l8ecm1uh dictator    Introduction
#> 48          ia9rnfvc     qyu3qg0q   survey    Demographics
#> 49          ia9rnfvc     qyu3qg0q   survey    Demographics
#> 50          ia9rnfvc     qyu3qg0q   survey    Demographics
#> 51          ia9rnfvc     qyu3qg0q   survey    Demographics
#> 52          iagvtslv     t0rog7nz   survey    Demographics
#> 53          iagvtslv     t0rog7nz   survey    Demographics
#> 54          iagvtslv     t0rog7nz   survey    Demographics
#> 55          iagvtslv     t0rog7nz   survey    Demographics
#> 56          iay3dhkn     7bfqtokx   survey    Demographics
#> 57          iay3dhkn     7bfqtokx   survey    Demographics
#> 58          iay3dhkn     7bfqtokx   survey    Demographics
#> 59          iay3dhkn     7bfqtokx   survey    Demographics
#> 60          jxh15obl     t0rog7nz   survey           Offer
#> 61          jxh15obl     t0rog7nz   survey           Offer
#> 62          jxh15obl     t0rog7nz   survey           Offer
#> 63          jxh15obl     t0rog7nz   survey           Offer
#> 64          ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 65          ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 66          ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 67          ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 68          lsl3vbij     7bfqtokx   survey    Demographics
#> 69          lsl3vbij     7bfqtokx   survey    Demographics
#> 70          lsl3vbij     7bfqtokx   survey    Demographics
#> 71          lsl3vbij     7bfqtokx   survey    Demographics
#> 72          rvce7958     7bfqtokx   survey    Demographics
#> 73          rvce7958     7bfqtokx   survey    Demographics
#> 74          rvce7958     7bfqtokx   survey    Demographics
#> 75          rvce7958     7bfqtokx   survey    Demographics
#> 76          vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 77          vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 78          vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 79          vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 80          wk247s9w     t0rog7nz   survey    Demographics
#> 81          wk247s9w     t0rog7nz   survey    Demographics
#> 82          wk247s9w     t0rog7nz   survey    Demographics
#> 83          wk247s9w     t0rog7nz   survey    Demographics
#> 84          xkobdvuh     t0rog7nz   survey    Demographics
#> 85          xkobdvuh     t0rog7nz   survey    Demographics
#> 86          xkobdvuh     t0rog7nz   survey    Demographics
#> 87          xkobdvuh     t0rog7nz   survey    Demographics
#> 88          xmxl46rm     vd1h01iv   survey    Demographics
#> 89          xmxl46rm     vd1h01iv   survey    Demographics
#> 90          xmxl46rm     vd1h01iv   survey    Demographics
#> 91          xmxl46rm     vd1h01iv   survey    Demographics
#> 92          xx78b3x0     vd1h01iv   survey    Demographics
#> 93          xx78b3x0     vd1h01iv   survey    Demographics
#> 94          xx78b3x0     vd1h01iv   survey    Demographics
#> 95          xx78b3x0     vd1h01iv   survey    Demographics
#> 96          zdki96d6     gii59rga   survey    Demographics
#> 97          znri6myc     l8ecm1uh dictator    Introduction
#> 98          znri6myc     l8ecm1uh dictator    Introduction
#> 99          znri6myc     l8ecm1uh dictator    Introduction
#> 100         znri6myc     l8ecm1uh dictator    Introduction
#>                                                  reason
#> 1         Experiment not completed. Noticed at: chatapp
#> 2   Experiment not completed. Noticed at: all_apps_wide
#> 3        Experiment not completed. Noticed at: dictator
#> 4          Experiment not completed. Noticed at: survey
#> 5   Experiment not completed. Noticed at: all_apps_wide
#> 6         Experiment not completed. Noticed at: chatapp
#> 7   Experiment not completed. Noticed at: all_apps_wide
#> 8        Experiment not completed. Noticed at: dictator
#> 9          Experiment not completed. Noticed at: survey
#> 10         Experiment not completed. Noticed at: survey
#> 11  Experiment not completed. Noticed at: all_apps_wide
#> 12       Experiment not completed. Noticed at: dictator
#> 13        Experiment not completed. Noticed at: chatapp
#> 14        Experiment not completed. Noticed at: chatapp
#> 15       Experiment not completed. Noticed at: dictator
#> 16         Experiment not completed. Noticed at: survey
#> 17  Experiment not completed. Noticed at: all_apps_wide
#> 18  Experiment not completed. Noticed at: all_apps_wide
#> 19       Experiment not completed. Noticed at: dictator
#> 20        Experiment not completed. Noticed at: chatapp
#> 21         Experiment not completed. Noticed at: survey
#> 22        Experiment not completed. Noticed at: chatapp
#> 23         Experiment not completed. Noticed at: survey
#> 24  Experiment not completed. Noticed at: all_apps_wide
#> 25       Experiment not completed. Noticed at: dictator
#> 26  Experiment not completed. Noticed at: all_apps_wide
#> 27       Experiment not completed. Noticed at: dictator
#> 28        Experiment not completed. Noticed at: chatapp
#> 29  Experiment not completed. Noticed at: all_apps_wide
#> 30         Experiment not completed. Noticed at: survey
#> 31        Experiment not completed. Noticed at: chatapp
#> 32  Experiment not completed. Noticed at: all_apps_wide
#> 33       Experiment not completed. Noticed at: dictator
#> 34         Experiment not completed. Noticed at: survey
#> 35         Experiment not completed. Noticed at: survey
#> 36  Experiment not completed. Noticed at: all_apps_wide
#> 37       Experiment not completed. Noticed at: dictator
#> 38        Experiment not completed. Noticed at: chatapp
#> 39        Experiment not completed. Noticed at: chatapp
#> 40  Experiment not completed. Noticed at: all_apps_wide
#> 41       Experiment not completed. Noticed at: dictator
#> 42         Experiment not completed. Noticed at: survey
#> 43  Experiment not completed. Noticed at: all_apps_wide
#> 44        Experiment not completed. Noticed at: chatapp
#> 45  Experiment not completed. Noticed at: all_apps_wide
#> 46       Experiment not completed. Noticed at: dictator
#> 47         Experiment not completed. Noticed at: survey
#> 48  Experiment not completed. Noticed at: all_apps_wide
#> 49       Experiment not completed. Noticed at: dictator
#> 50        Experiment not completed. Noticed at: chatapp
#> 51         Experiment not completed. Noticed at: survey
#> 52        Experiment not completed. Noticed at: chatapp
#> 53         Experiment not completed. Noticed at: survey
#> 54  Experiment not completed. Noticed at: all_apps_wide
#> 55       Experiment not completed. Noticed at: dictator
#> 56         Experiment not completed. Noticed at: survey
#> 57  Experiment not completed. Noticed at: all_apps_wide
#> 58       Experiment not completed. Noticed at: dictator
#> 59        Experiment not completed. Noticed at: chatapp
#> 60  Experiment not completed. Noticed at: all_apps_wide
#> 61       Experiment not completed. Noticed at: dictator
#> 62        Experiment not completed. Noticed at: chatapp
#> 63         Experiment not completed. Noticed at: survey
#> 64        Experiment not completed. Noticed at: chatapp
#> 65  Experiment not completed. Noticed at: all_apps_wide
#> 66       Experiment not completed. Noticed at: dictator
#> 67         Experiment not completed. Noticed at: survey
#> 68        Experiment not completed. Noticed at: chatapp
#> 69       Experiment not completed. Noticed at: dictator
#> 70  Experiment not completed. Noticed at: all_apps_wide
#> 71         Experiment not completed. Noticed at: survey
#> 72        Experiment not completed. Noticed at: chatapp
#> 73  Experiment not completed. Noticed at: all_apps_wide
#> 74       Experiment not completed. Noticed at: dictator
#> 75         Experiment not completed. Noticed at: survey
#> 76  Experiment not completed. Noticed at: all_apps_wide
#> 77       Experiment not completed. Noticed at: dictator
#> 78        Experiment not completed. Noticed at: chatapp
#> 79         Experiment not completed. Noticed at: survey
#> 80        Experiment not completed. Noticed at: chatapp
#> 81         Experiment not completed. Noticed at: survey
#> 82  Experiment not completed. Noticed at: all_apps_wide
#> 83       Experiment not completed. Noticed at: dictator
#> 84        Experiment not completed. Noticed at: chatapp
#> 85       Experiment not completed. Noticed at: dictator
#> 86         Experiment not completed. Noticed at: survey
#> 87  Experiment not completed. Noticed at: all_apps_wide
#> 88       Experiment not completed. Noticed at: dictator
#> 89        Experiment not completed. Noticed at: chatapp
#> 90  Experiment not completed. Noticed at: all_apps_wide
#> 91         Experiment not completed. Noticed at: survey
#> 92        Experiment not completed. Noticed at: chatapp
#> 93  Experiment not completed. Noticed at: all_apps_wide
#> 94       Experiment not completed. Noticed at: dictator
#> 95         Experiment not completed. Noticed at: survey
#> 96  Experiment not completed. Noticed at: all_apps_wide
#> 97        Experiment not completed. Noticed at: chatapp
#> 98       Experiment not completed. Noticed at: dictator
#> 99  Experiment not completed. Noticed at: all_apps_wide
#> 100        Experiment not completed. Noticed at: survey
#> 
#> $unique
#>    participant.code session.code  end_app        end_page
#> 1          1l5kal0r     t0rog7nz   survey    Demographics
#> 5          1mn1kzs7     gii59rga   survey    Demographics
#> 6          2d72mfgh     7bfqtokx   survey    Demographics
#> 10         2scvem7a     vd1h01iv dictator ResultsWaitPage
#> 14         3ttf7yix     7bfqtokx   survey    Demographics
#> 18         46kxib6w     7bfqtokx   survey    Demographics
#> 22         4zhzdmzo     7bfqtokx   survey    Demographics
#> 26         6cim3a5i     gii59rga   survey    Demographics
#> 27         7l0hmpcq     qyu3qg0q   survey    Demographics
#> 31         7wa8kk3d     l8ecm1uh dictator    Introduction
#> 35         a7dppel1     t0rog7nz   survey    Demographics
#> 39         ao9kqvqn     l8ecm1uh dictator    Introduction
#> 43         byxco6zo     gii59rga   survey    Demographics
#> 44         d4sq7zio     l8ecm1uh dictator    Introduction
#> 48         ia9rnfvc     qyu3qg0q   survey    Demographics
#> 52         iagvtslv     t0rog7nz   survey    Demographics
#> 56         iay3dhkn     7bfqtokx   survey    Demographics
#> 60         jxh15obl     t0rog7nz   survey           Offer
#> 64         ktjz5jli     7bfqtokx dictator ResultsWaitPage
#> 68         lsl3vbij     7bfqtokx   survey    Demographics
#> 72         rvce7958     7bfqtokx   survey    Demographics
#> 76         vbhvhozv     vd1h01iv dictator ResultsWaitPage
#> 80         wk247s9w     t0rog7nz   survey    Demographics
#> 84         xkobdvuh     t0rog7nz   survey    Demographics
#> 88         xmxl46rm     vd1h01iv   survey    Demographics
#> 92         xx78b3x0     vd1h01iv   survey    Demographics
#> 96         zdki96d6     gii59rga   survey    Demographics
#> 97         znri6myc     l8ecm1uh dictator    Introduction
#>                      reason
#> 1  Experiment not completed
#> 5  Experiment not completed
#> 6  Experiment not completed
#> 10 Experiment not completed
#> 14 Experiment not completed
#> 18 Experiment not completed
#> 22 Experiment not completed
#> 26 Experiment not completed
#> 27 Experiment not completed
#> 31 Experiment not completed
#> 35 Experiment not completed
#> 39 Experiment not completed
#> 43 Experiment not completed
#> 44 Experiment not completed
#> 48 Experiment not completed
#> 52 Experiment not completed
#> 56 Experiment not completed
#> 60 Experiment not completed
#> 64 Experiment not completed
#> 68 Experiment not completed
#> 72 Experiment not completed
#> 76 Experiment not completed
#> 80 Experiment not completed
#> 84 Experiment not completed
#> 88 Experiment not completed
#> 92 Experiment not completed
#> 96 Experiment not completed
#> 97 Experiment not completed
#> 
#> $all_end
#>           
#>            Demographics Introduction Offer ResultsWaitPage chat
#>   chatapp             0            0     0               0    1
#>   dictator            0            4     0               3    0
#>   survey             20            0     1               0    0
#> 
#> $codes
#>  [1] "1l5kal0r" "1mn1kzs7" "2d72mfgh" "2scvem7a" "3ttf7yix" "46kxib6w"
#>  [7] "4zhzdmzo" "6cim3a5i" "7l0hmpcq" "7wa8kk3d" "a7dppel1" "ao9kqvqn"
#> [13] "byxco6zo" "d4sq7zio" "ia9rnfvc" "iagvtslv" "iay3dhkn" "jxh15obl"
#> [19] "ktjz5jli" "lsl3vbij" "rvce7958" "vbhvhozv" "wk247s9w" "xkobdvuh"
#> [25] "xmxl46rm" "xx78b3x0" "zdki96d6" "znri6myc"
#> 
#> $count
#> [1] 28
#>